internal-links: open non-img urls in hidden webview.
This will make sure that the current server webview will
not reload due to url change.
